requirements.txt
biopython==1.85
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
diff-match-patch==20241021
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.3
pycparser==2.22
python-dotenv==1.1.0
Werkzeug==3.1.3
